Renamed test_components to components and adjusted relevant function call in dashboard.py
