+++ components/proto-img-only.py
