+++ components/prototype_histogram.py
