+++ dashboard.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import plotly.express as px
 from dash import Dash, html, dcc, Input, Output, State
-from test_components.query import get_species_options
+from components.query import get_species_options
 
 IMAGES_BASE_URL = "https://github.com/Imageomics/dashboard-prototype/blob/main/test_data/images/"
